% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/z5_count_obs_sr_pairs_function.R
\name{count_obs_sr_pairs}
\alias{count_obs_sr_pairs}
\title{Count the number of observed SR pairs}
\usage{
count_obs_sr_pairs(x, a_max, na)
}
\arguments{
\item{x}{Numeric vector consisting of 0 (not-observed) and 1 (observed) values
with length equal to the number of possibly observable years}

\item{a_max}{Numeric vector of length 1: the maximum age at which fish can return}

\item{na}{Numeric vector of length 1: the ages at which fish can return}
}
\description{
Count the number of observed SR pairs
}
