% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/z1_eq_ricker_function.R
\name{eq_ricker}
\alias{eq_ricker}
\title{Calculate Equilibrium Quantities from a Ricker Function}
\usage{
eq_ricker(alpha, beta, U_msy, S_msy, U_range)
}
\arguments{
\item{alpha}{a numeric vector of length 1.}

\item{beta}{a numeric vector of length 1.}

\item{U_msy}{a numeric vector of length 1.}

\item{S_msy}{a numeric vector of length 1.}

\item{U_range}{a numeric vector storing the exploitation
rates to calculate the equilibrium quantities at.}
}
\description{
Given known (or estimated) values of alpha, beta,
  U_msy, and S_msy, for a single stock,
  calculate Seq and Ceq at each of a set of exploitation rates
}
