% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/z8_example_sim.R
\name{example_sim}
\alias{example_sim}
\title{An example simulation wrapper}
\usage{
example_sim(do_plot = F, save = F)
}
\arguments{
\item{do_plot}{logical. Do you want to make an SRA plot?}

\item{save}{logical. Do you want to save a file containing the plot?
Only takes effect if \code{do_plot = TRUE}}

\item{U_SUM}{passed to \code{init_sim()}}

\item{min_sigR}{passed to \code{init_sim()}}

\item{max_sigR}{passed to \code{init_sim()}}

\item{max_p_overfished}{passed to \code{init_sim()}}

\item{rho}{passed to \code{init_sim()}}
}
\description{
An example simulation wrapper
}
\examples{
example_sim(
  do_plot = F,
  save = F
  )

}
