% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/z3_gen_lm_mgmt.R
\name{gen_lm_mgmt}
\alias{gen_lm_mgmt}
\title{Obtain estimates of U_msy and S_msy given alpha and beta}
\usage{
gen_lm_mgmt(alpha, beta)
}
\arguments{
\item{@alpha}{a numeric vector representing the alpha parameter from a SRA.
Can be length > 1.}

\item{@beta}{a numeric vector representing the beta parameter from a SRA.
Can be length > 1.}
}
\description{
Converts alpha and beta into Smsy and Umsy.
}
\note{
The conversion of alpha to U_msy is an approximation, as there is no analytical solution.
}
