% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/z2_gen_mgmt_function.R
\name{gen_mgmt}
\alias{gen_mgmt}
\title{Generate Quantities used in Management}
\usage{
gen_mgmt(params)
}
\arguments{
\item{params}{A list created using \code{init_sim()}.}
}
\description{
Based on vectors of substock-specific parameters, obtain the biological reference points
  for the aggregate stock if all substocks were fished at the same rate.
}
