% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3c_gen_R_ys_obs_function.R
\name{gen_Rys_obs}
\alias{gen_Rys_obs}
\title{Obtain Brood Year Recruits}
\usage{
gen_Rys_obs(params, obs)
}
\arguments{
\item{params}{A list created using \code{\link{init_sim}}.}

\item{obs}{A list created using \code{\link{obs_sim}}.}
}
\value{
An updated list which contains all of the contents of the \code{obs} list
  supplied as an argument
}
\description{
Reconstruct the brood table to get the number of fish that were produced by a single
spawning event. For indexing, recruit abundance in brood year \code{a_max+1}
was produced by spawner abundance in calendar year 1.
}
\details{
Currently, a brood year recruitment event becomes an NA if any
  one of the return years was missing in the escapement counts.
  For stocks without age composition sampling, the average
  age composition observed for monitored stocks is used to allocate
  abundance-at-age-and-year to brood year recruits. The assumption is made that
  all substocks received the same exploitation rate.
}
