% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/z4_implement_error_function.R
\name{implement_error}
\alias{implement_error}
\title{Generate Management Errors}
\usage{
implement_error(U_target, SUM)
}
\arguments{
\item{U_target}{Numeric vector of length 1: the mean target exploitation rate}

\item{SUM}{Numeric vector of length 1: the sum of the two shape parameters of the beta distribution.
The larger this value is, the less implementation error there is.}
}
\description{
Generate a beta random deviate around a target exploitation rate
with variability given by \code{SUM}, the sum of the two beta shape parameters.
}
