% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/4_params_summary_function.R
\name{input_summary}
\alias{input_summary}
\title{Summarize important input quantities}
\usage{
input_summary(params, obs, true, seed)
}
\arguments{
\item{params}{a list object created with \code{init_sim()}}

\item{obs}{a list object created with the observation generation
functions workflow.}

\item{true}{a list object created by \code{ricker_sim()}}

\item{seed}{An identifier for this set of inputs}
}
\description{
Reformats key information from the
  three primary input lists for writing to output files
}
