% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/4_params_summary_function.R
\name{params_summary}
\alias{params_summary}
\title{Summarize true parameters}
\usage{
params_summary(params, obs, seed)
}
\arguments{
\item{params}{a list object created with \code{init_sim()}}

\item{obs}{a list object created with the observation generation
functions workflow.}

\item{seed}{An identifier for this set of parameters}
}
\description{
Reformats key output from the \code{params} and \code{obs}
lists for writing to output files.
}
