% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/z6_prob2pi_function.R
\name{pi2beta}
\alias{pi2beta}
\title{Maturity Conversion Function 3}
\usage{
pi2beta(pi)
}
\description{
Converts from \code{pi} to a set of logit-scale coefficients that is used to obtain
the \code{prob} vector
}
