% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/z6_prob2pi_function.R
\name{pi2prob}
\alias{pi2prob}
\title{Maturity Conversion Function 1}
\usage{
pi2prob(pi)
}
\description{
Converts a vector from the marginal probability of maturity at each age
to a vector containing the conditional probabilities of maturity at ages
\code{a:(a_max-1)} given not yet matured.
}
