% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2_pop_sim_function.R
\name{ricker_sim}
\alias{ricker_sim}
\title{Generate True States}
\usage{
ricker_sim(params, nb = 50, U_strategy = "Ustar_0.3",
  stock_sd = 0.25, year_sd = 0.2, stock_year_rho = 0.9)
}
\arguments{
\item{params}{A list created using \code{init_sim()}.}

\item{nb}{Numeric vector of length 1: the number of years to simulate the population
before tracking the states of interest.}

\item{U_strategy}{Character vector of length 1. Specifies the way the harvest rate target
is selected based on the true biological reference points. Valid options currently are:
\code{"Ustar_0.1"}, \code{"Ustar_0.3"}, \code{"Ustar_0.5"}, \code{"U_MSY"}. \code{"Ustar_0.1"} is
defined as the exploitation rate applied equally to all substocks that would result in
no more than 10\% being overfished.}

\item{stock_sd}{numeric vector of length 1: the standard deviation of stock effects in the maturity schedule formation}

\item{year_sd}{numeric vector of length 1: the standard deviation of year effects in the maturity schedule formation}

\item{stock_year_rho}{numeric vector of length 1: the correlation between stocks and years in the maturity schedule formation}
}
\value{
A list with the true states simulated.
}
\description{
Simulates Ricker dynamics based on the settings supplied via the input argument.
}
