% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/z7_true_plot.R
\name{true_plot}
\alias{true_plot}
\title{Create a plot of the true states}
\usage{
true_plot(params, true, new_window = T, type = "SRA")
}
\arguments{
\item{params}{A list created with \code{init_sim()}}

\item{true}{A list created with \code{ricker_sim()}}

\item{new_window}{logical. Do you want the plots to open in a new window?}

\item{type}{character vector. Current accepted types are \code{"SRA"} and \code{"spawners"}}
}
\description{
Create a plot of the true states
}
