% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/z2_gen_mgmt_function.R
\name{gen_mgmt}
\alias{gen_mgmt}
\title{Generate Quantities used in Management}
\usage{
gen_mgmt(params, U_range = seq(0, 1, 0.01))
}
\arguments{
\item{params}{A list created using \code{\link{init_sim}}.}

\item{U_range}{A sequence at which to calculate the different equilibrium states}
}
\description{
Based on vectors of substock-specific parameters, obtain the biological reference points
  for the aggregate stock if all substocks were fished at the same rate. Calculates
  a set of quantities called \code{Sstar_p} and \code{Ustar_p}. \code{Sstar_p}
  is the smallest aggregate escapement where you can still have no more than p*100\% of
  the substocks not overfished. \code{Ustar_p} is the maximum exploitation rate without more
  than p*100\% of the substocks overfished.
}
