% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1_gen_params_function.R
\name{init_sim}
\alias{init_sim}
\title{Initialize Simulation}
\usage{
init_sim(nt = 42, a_min = 4, a_max = 7, U_msy, S_msy, rho = 0.5,
  U_SUM = 100, max_p_overfished = 0.3, min_S_cv = 0.1,
  max_S_cv = 0.2, min_C_cv = 0.1, max_C_cv = 0.2, x_ESS = 100)
}
\arguments{
\item{nt}{Numeric vector of length 1: the number of
calendar years with observations to simulate}

\item{a_min}{Numeric vector of length 1: youngest age at maturity}

\item{a_max}{Numeric vector of length 1: oldest age at maturity}

\item{U_msy}{Numeric vector storing U_msy for each stock to simulate}

\item{S_msy}{Same as \code{S_msy}, except S_msy. Must be the same
length as \code{U_msy}}

\item{U_SUM}{Numeric vector of length 1: beta sample size of implementation error}

\item{max_p_overfished}{Numeric vector of length 1: used in setting the constant exploitation rate}

\item{min_S_cv}{Numeric vector of length 1: smallest observation CV for any substocks's escapement.}

\item{max_S_cv}{Numeric vector of length 1: largest observation CV for any substocks's escapement.}

\item{min_C_cv}{Numeric vector of length 1: smallest observation CV for any year's total harvest.}

\item{max_C_cv}{Numeric vector of length 1: largest observation CV for any year's total harvest.}

\item{x_ESS}{Numeric vector of length 1: the effective sample size for scale sampling on stocks that have age data.}
}
\description{
Handles setting up the dimensional variables
and drawing random quantities for driving the
operating model and observation
}
\details{
A covariance matrix for the recruitment residuals \code{Sigma} is generated by this function.
  It creates it with an inverse wishart distribution, such that \code{sqrt(diag(Sigma))} elements are
  between 0.2 and 0.5 (mean around 0.3) and the correlation component of the covariance terms is between
  roughly -0.5 and 0.5 with a mean of approximately 0.1.
}
