% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3b_obs_filter_function.R
\name{obs_filter}
\alias{obs_filter}
\title{Sample Observed States}
\usage{
obs_filter(params, obs, mimic = NULL, minSRobs = 3)
}
\arguments{
\item{params}{A list created using \code{init_sim()}.}

\item{obs}{A list created using \code{obs_sim()}.}

\item{mimic}{A data frame with columns \code{"stock"} (a stock identifier)
and \code{"obs"} (a binary indicator for whether that stock was monitored that year).
This function attempts to mimic this pattern of observation frequency. If \code{NULL} (default),
the default is for each year to have a 60% of being monitored for each stock.}

\item{minSRobs}{Numeric vector of length 1: the minimum number of fully observed
spawner and recruit brood year pairs allowed. Defaults to 3.}
}
\value{
A list containing the filtered observed states.
  Any year not sampled will be an NA.
}
\description{
Based on the observed states obtained using \code{obs_sim},
introduce a sampling frequency.
}
