% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3b_obs_filter_function.R
\name{obs_filter}
\alias{obs_filter}
\title{Sample Observed States}
\usage{
obs_filter(params, obs)
}
\arguments{
\item{params}{A list created using \code{init_sim()}.}

\item{obs}{A list created using \code{obs_sim()}.}
}
\value{
A list containing the filtered observed states.
  Any year not sampled will be an NA.
}
\description{
Based on the observed states obtained using \code{obs_sim},
introduce a sampling frequency.
}
