% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3a_obs_sim_function.R
\name{obs_sim}
\alias{obs_sim}
\title{Generate Observed States}
\usage{
obs_sim(params, true)
}
\arguments{
\item{params}{A list created using \code{\link{init_sim}}.}

\item{true}{A list created using \code{\link{ricker_sim}}.}
}
\value{
A list containing the observed states,
  which are not necessarily in a format ready for model-fitting
}
\description{
Based on the true states obtained using \code{\link{ricker_sim}}, introduce observation errors
for all possibly observable calendar years
}
