% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2_pop_sim_function.R
\name{ricker_sim}
\alias{ricker_sim}
\title{Generate True States}
\usage{
ricker_sim(params)
}
\arguments{
\item{params}{A list created using \code{init_sim()}.}
}
\value{
A list with the true states simulated.
}
\description{
Simulates Ricker dynamics based on the settings supplied via the input argument.
}
