% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/z7_true_plot.R
\name{true_plot}
\alias{true_plot}
\title{Create a plot of the true states}
\usage{
true_plot(params, true, ext_device = T, type = "SRA")
}
\arguments{
\item{params}{A list created with \code{\link{init_sim}}}

\item{true}{A list created with \code{\link{ricker_sim}}}

\item{ext_device}{logical. Do you want the plots to open in an external device?}

\item{type}{character vector. Current accepted types are \code{"SRA"} and \code{"spawners"}}
}
\description{
Create a plot of the true states
}
